Change into english and fix responsive
