--- src/utils/api.js
@@ -1,163 +1,0 @@
-const flights = [
-  {
-    id: 1,
-    airline: "Jet Air",
-    class: "Economy",
-    departureTime: "07:00",
-    arrivalTime: "12:00",
-    totalTime: "4h 0m",
-    type: "Direct",
-    departureAirportId: "JKT",
-    arrivalAirportId: "MLB",
-    price: "IDR 4.950.000",
-    date: "3 Maret 2023",
-    dep_airport: "Soekarno Hatta - Terminal 1A Domestik",
-    code: "JT - 203",
-    arr_airport: "Melbourne International Airport",
-  },
-  {
-    id: 2,
-    airline: "AirAsia",
-    class: "Economy",
-    departureTime: "09:00",
-    arrivalTime: "14:00",
-    totalTime: "5h 0m",
-    type: "Direct",
-    departureAirportId: "JKT",
-    arrivalAirportId: "MLB",
-    price: "IDR 5.200.000",
-    date: "4 Maret 2023",
-    dep_airport: "Soekarno Hatta - Terminal 1B Domestik",
-    code: "AA - 101",
-    arr_airport: "Melbourne International Airport",
-  },
-  {
-    id: 3,
-    airline: "Garuda Indonesia",
-    class: "Business",
-    departureTime: "11:00",
-    arrivalTime: "16:00",
-    totalTime: "5h 0m",
-    type: "Direct",
-    departureAirportId: "JKT",
-    arrivalAirportId: "MLB",
-    price: "IDR 10.500.000",
-    date: "5 Maret 2023",
-    dep_airport: "Soekarno Hatta - Terminal 2F Internasional",
-    code: "GA - 305",
-    arr_airport: "Melbourne International Airport",
-  },
-  {
-    id: 4,
-    airline: "Lion Air",
-    class: "Economy",
-    departureTime: "06:00",
-    arrivalTime: "11:00",
-    totalTime: "5h 0m",
-    type: "Direct",
-    departureAirportId: "JKT",
-    arrivalAirportId: "MLB",
-    price: "IDR 4.800.000",
-    date: "6 Maret 2023",
-    dep_airport: "Soekarno Hatta - Terminal 1C Domestik",
-    code: "LI - 402",
-    arr_airport: "Melbourne International Airport",
-  },
-  {
-    id: 5,
-    airline: "Citilink",
-    class: "Economy",
-    departureTime: "08:00",
-    arrivalTime: "13:00",
-    totalTime: "5h 0m",
-    type: "Direct",
-    departureAirportId: "JKT",
-    arrivalAirportId: "MLB",
-    price: "IDR 4.600.000",
-    date: "7 Maret 2023",
-    dep_airport: "Soekarno Hatta - Terminal 1A Domestik",
-    code: "CI - 206",
-    arr_airport: "Melbourne International Airport",
-  },
-  {
-    id: 6,
-    airline: "Sriwijaya Air",
-    class: "Economy",
-    departureTime: "07:30",
-    arrivalTime: "12:30",
-    totalTime: "5h 0m",
-    type: "Direct",
-    departureAirportId: "JKT",
-    arrivalAirportId: "MLB",
-    price: "IDR 5.000.000",
-    date: "8 Maret 2023",
-    dep_airport: "Soekarno Hatta - Terminal 1B Domestik",
-    code: "SJ - 501",
-    arr_airport: "Melbourne International Airport",
-  },
-  {
-    id: 7,
-    airline: "Batik Air",
-    class: "Economy",
-    departureTime: "10:00",
-    arrivalTime: "15:00",
-    totalTime: "5h 0m",
-    type: "Direct",
-    departureAirportId: "JKT",
-    arrivalAirportId: "MLB",
-    price: "IDR 5.300.000",
-    date: "9 Maret 2023",
-    dep_airport: "Soekarno Hatta - Terminal 2E Internasional",
-    code: "BI - 303",
-    arr_airport: "Melbourne International Airport",
-  },
-  {
-    id: 8,
-    airline: "Qantas",
-    class: "Business",
-    departureTime: "12:00",
-    arrivalTime: "17:00",
-    totalTime: "5h 0m",
-    type: "Direct",
-    departureAirportId: "JKT",
-    arrivalAirportId: "MLB",
-    price: "IDR 12.000.000",
-    date: "10 Maret 2023",
-    dep_airport: "Soekarno Hatta - Terminal 2D Internasional",
-    code: "QA - 602",
-    arr_airport: "Melbourne International Airport",
-  },
-  {
-    id: 9,
-    airline: "Singapore Airlines",
-    class: "Economy",
-    departureTime: "13:00",
-    arrivalTime: "18:00",
-    totalTime: "5h 0m",
-    type: "Direct",
-    departureAirportId: "JKT",
-    arrivalAirportId: "MLB",
-    price: "IDR 5.700.000",
-    date: "11 Maret 2023",
-    dep_airport: "Soekarno Hatta - Terminal 2C Internasional",
-    code: "SA - 703",
-    arr_airport: "Melbourne International Airport",
-  },
-  {
-    id: 10,
-    airline: "Emirates",
-    class: "First Class",
-    departureTime: "14:00",
-    arrivalTime: "19:00",
-    totalTime: "5h 0m",
-    type: "Direct",
-    departureAirportId: "JKT",
-    arrivalAirportId: "MLB",
-    price: "IDR 20.000.000",
-    date: "12 Maret 2023",
-    dep_airport: "Soekarno Hatta - Terminal 3 Internasional",
-    code: "EK - 804",
-    arr_airport: "Melbourne International Airport",
-  },
-];
-export default flights;
